# Speed up some draw bounds tests