This update addresses the comment/issue mentioned by @MightyJosip in issue #1212 that the `test_arc__bounding_rect()` test is performing a lot of iterations and taking much longer than the other draw test cases. Changes were also made to some of the other draw bounds tests (ellipse/polygon/rect) to reduce their test times.

---

Overview of changes:
- Reduced the size of the surfaces being checked
- Reduced the number of iterations on some of the test parameters

---

Test details:
- ran three passes to average out times
- baseline
  - python: 3.8.1 (64bit)
    - cmd: python -Wd -m pygame.tests draw_test
      - Ran 205 tests in 57.065s
      - Ran 205 tests in 50.898s
      - Ran 205 tests in 50.937s
- with changes to bounding tests
  - python: 3.8.1 (64bit)
    - cmd: python -Wd -m pygame.tests draw_test
      - Ran 205 tests in 5.777s
      - Ran 205 tests in 5.869s
      - Ran 205 tests in 5.565s
  - python: 2.7.10 (64bit)
      - cmd: python -Wd -m pygame.tests draw_test
        - Ran 205 tests in 8.432s
        - Ran 205 tests in 6.878s
        - Ran 205 tests in 7.097s

---

System details:
- os: windows 10 (64bit)
- python: 3.8.1 (64bit) and 2.7.10 (64bit)
- pygame: 2.0.0.dev7 (SDL: 2.0.10) at 2ff089bfcf7fffc7f8a566a3cc3d0d1fb76fe9e1