@MightyJosip 2020-01-30
Thanks, in my opinion it looks great